# backend/ibpms-core/src/main/java/com/ibpms/core/ai/controller/CognitiveAiController.java
package com.ibpms.core.ai.controller;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.Map;

/**
 * REST Controller for Cognitive AI operations (Mock implementation for
 * Integration Gaps).
 */
@RestController
@RequestMapping("/api/v1/ai")
public class CognitiveAiController {

    @PostMapping("/correct")
    public ResponseEntity<Map<String, String>> correctText(@RequestBody Map<String, Object> request) {
        // Mock Implementation
        String originalText = (String) request.getOrDefault("original", "");
        String prompt = (String) request.getOrDefault("prompt", "Mejorar redacción");

        String correctedText = "[IA Mejorado (" + prompt + ")]: " + originalText;

        return ResponseEntity.ok(Map.of(
                "correctedText", correctedText,
                "status", "SUCCESS"));
    }

    @PostMapping("/dmn/translate")
    public ResponseEntity<Map<String, Object>> translateToDmn(@RequestBody Map<String, Object> request) {
        // Mock Implementation for DMN Translator
        String requirements = (String) request.getOrDefault("requirements", "");

        return ResponseEntity.ok(Map.of(
                "dmnXml",
                "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<definitions xmlns=\"https://www.omg.org/spec/DMN/20191111/MODEL/\" id=\"Definitions_Mock\">\n  <!-- Mock DMN generated for: "
                        + requirements + " -->\n</definitions>",
                "confidenceScore", 95,
                "status", "GENERATED"));
    }
}