# docs/api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: iBPMS Platform Core API (PoC)
  description: API Central para la Plataforma de Automatización Inteligente (iBPMS + Case Management). Expone los Driving Adapters del Hexágono.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de Desarrollo Local (Spring Boot)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /ai/correct:
    post:
      summary: Corrección de texto con Inteligencia Artificial
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                original:
                  type: string
                prompt:
                  type: string
      responses:
        '200':
          description: Texto corregido con IA
...
  /ai/dmn/translate:
    post:
      summary: NLP a DMN (Traducir Lenguaje Natural a Reglas de Negocio XML)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: XML Generado
...